requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==2.0.12
Django==4.2.21
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2025.2
requests==2.26.0
sqlparse==0.5.3
toml==0.10.2
typing_extensions==4.13.2
urllib3==1.26.20
